Extract TeamCity-specific helpers to a trait
